add non-file backed dataset ingest
If datasets.json doesn't define a "publish" block for the dataset,
the dataset directory is not copied anywhere. The metadata is still
ingested into GRQ.
If datasets.json doesn't define a "browse" block for the dataset,
the browse images are not copied anywhere. The metadata is still
ingested into GRQ.
Both "publish" and "browse" blocks are independent.